refactor: split doc_utils into multiple files
